Revert "fbcon: use fbconsole mode to init new consoles"
This reverts commit e19dc11666087f7090b79d91eead2125d928c7ec.

This change was originally intended to help fbconsole retain its
original mode when a graphical client like X gets killed abruptly, and
the mode which was chosen by X would get forcefully programmed to the
console.

But, there are cases where it is better to update fbconsole's mode
instead of sticking to its original mode because it might be invalid.
For instance, if a new monitor is plugged-in when X is active, the old
modelist is purged and a new one is created. Fbconsole's mode too is
updated as part of hotplug worker (currently, first mode from the
modelist is chosen). When X is killed and fbconsole takes over, if the
original console mode isn't available in the new modelist, fbconsole
would fail to initialize. Instead, its better to initialize fbconsole
with the new mode, which is guaranteed to be available in the new
modelist.

Also, tegra framebuffer driver now takes care of disabling display head
before updating with a new mode. Hence, the inconsistencies seen in
fbconsole's mode update during suspend/resume (noted in original patch)
are no more seen.

bug 200453236

Change-Id: I6cbd61ecfe9c6dea6e85acf8d059c0a9f126c37e
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1954691
(cherry picked from commit 30ffbb281c7c2d9638b587fc281c987ed376943a)
Reviewed-on: https://git-master.nvidia.com/r/1959011
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>