video: tegra: avoid divide-by-zero during boot
gcc 7.3 detects a code-path that can perform a divide-by-zero and replaces
this with an explicit "brk #0x3e8" (i.e. "brk #1000") instruction, which
causes a crash if the code is executed.
Adding condition to check non-zero value in code removes brk statement.
Changes are done to make sure, it does not divide-by-zero during compilation.

Bug 200442298

Change-Id: I83a7119167a17d65d4e141e24dbd01c08ef27438
Signed-off-by: Shubhi Garg <shgarg@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1959377
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sandipan Patra <spatra@nvidia.com>
Tested-by: Sandipan Patra <spatra@nvidia.com>
Reviewed-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>