platform: tegra: mcerr-t19x: fix mc-err print that is showing additional number
Fix mc-err print that has additional number for hi_addr_reg printed
and misleading the fault address to be treated as > 32-bit even when
it is not.
On T19x, addr printed is already 64-bit in the mc-err print. Make
it explicit by changing the format of addr printed.

Bug 200431631

Change-Id: I69c3d74616b8b005ed707c625d3dcf594e5cf5f1
Signed-off-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1965522
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>