video: nvmap: Return Error on reading VPR
On T186 and above, reading VPR resulted in reboots
when accessed through CPU(as this is not allowed).

Because of above, a special condition that was introduced to
return an error with filling the buffer with 0xff(for
backward compatibility).

The proposal is to simply return an error(No Permission error) without
reading or emulating by writing 0xFF to the buffer passed for reading.

This changes the implementation for t210, but provides an unified
behaviour across chips.

Bug 2459437

Change-Id: Iafc281541c4a5dd65e2337baec689ad4bfc81e05
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1969147
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>