PCI: fix NV test endpoint function
The Tegra instantiation of the DWC PCIe controller implements a minimum
ATU entry size of 64K. Adapt the test "function" driver to use this size.
Without this change, some bits of the IOVA will be dropped when written
to the ATU lower target address register, which will cause transactions
to be directed to the wrong IOVA.

Throw in some unrelated message formatting fixes too.

Bug 2312059

Change-Id: I0700dc6193769ed295d5ba4641a2168f114aa464
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1969811
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>