dc: fix vrr timestamp wait overflow
Fix a signed 32-bit negative overflow, which occurred when timestamp_ns
is much less than now_ns

The symptom of this overflow is that animation runs at 1 FPS for about a
minute, rarely after a long uptime and an idle period.  I observed this
on ATV Homescreen and have reproduced a few times.

The overflow (negative) occurred here:
    sleep_us = (s32)div_s64(timestamp_ns - now_ns, 1000ll);

* timestamp_ns is passed through SF and HWC, after being provided by
applications, and is a CLOCK_MONOTONIC timestamp.
* now_ns is current timestamp from ktime_get_ts

Example observed values from an overflow are
* timestamp_ns = 53986436618520 // 900 minutes
* now_ns = 58068360329723 // 967 minutes
* div_s64(timestamp_ns - now_ns, 1000ll) = −4081923711
** This -4B is too large for s32, even after the div by 1000
* (s32)(−4081923711ll) = 213043585 // 213 seconds
** This gets ceilinged to a 1-second sleep each frame

When this issue reproduces, the timestamp_ns is not changing each frame,
like it is from a buffer where the timestamp is no longer getting
updated.  If the timestamp was set to 0 in userspace, an early out
branch would have been taken.

Jira TGFXP-2581
Bug 2329754

Change-Id: Ia22a175432e454f8c58fe4763f93172b7386cfcc
Signed-off-by: Jon McCaffrey <jmccaffrey@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1972566
(cherry picked from commit e670da15f4b4ab18e983f4004ba5b638f5ccc5c2)
Reviewed-on: https://git-master.nvidia.com/r/1987053
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>