gpu: nvgpu: rfr: Notify user of email destination
When sending emails it can be nice to see who you are sending an
email to if you specify multiple '--to' options.

Also update the get_user_message() function to remove comment lines
starting with '#'. This lets the script place information useful
to a developer in the message that won't ultimately make it to the
final email.

[Bump minor revision to 1.0.2]

Change-Id: I3657a787d8e9a8c19de727e050a0b9a18a6d43e0
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1983642
Reviewed-by: Rohit Khanna <rokhanna@nvidia.com>