tegra234: Introduce soc-minimal.dtsi
Introduce tegra234-soc-minimal.dtsi that is intended for just the
minimal nodes required to boot to shell. This is geared towards usage in
pre-silicon platform bringup.
Everytime a new platform like QT, Emulation, FPGA (in early stages)
is brought up, minimal nodes with boot to shell is the primary
requirement. Hence, each of the platform specific dtsi can inherit this
minimal.dtsi, add chosen nodes, add minimal overrides & this should serve
as the standard recipe for boot to shell.

This file is to be included in the base dtsi files & have overrides in
place to ensure that nodes aren't repeated for regular full fledgeed
operation like VDK or in future Silicon.

The nodes that are required are:
- Interrupt controller
- Memory
- Timer
- Serial port

JIRA TKER-242

Change-Id: I47821e6af70e943b05446280741ccfea29ab8adf
Signed-off-by: Vignesh Radhakrishnan <vigneshr@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1987909
GVS: Gerrit_Virtual_Submit
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>