ALSA: hda/tegra: Silence error if switch name is not found
The sysfs entry 'hda_pcm_map' is only used for Android to describe
the mapping between display interfaces and hda interfaces.

With L4T due to the ordering which the display and hda devices are
probed the following errors message are seen on boot on the various
Jetson platforms ...

 tegra-hda 3510000.hda: error in getting switch name for hda_pcm_id(3)

Even for Android the above error message is seen for any hda interfaces
that are not used on a given platform.

Ideally this should be fixed by ...
1. Disabling any unused hda interfaces in device-tree
2. Ensuring that the hda interfaces are probe deferred until all the
   displays are registered.

Deferring the probe of the hda devices is a non-trivial to fix and
given that these error messages are not critical, silence them for now
by making them a dev_dbg() message.

Bug 200469524

Change-Id: I1d1f5b4e88d7a6f296010466805b85e44a15ae23
Signed-off-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2000402
(cherry picked from commit 16d873be904463a2f837f54899cdcdce2f27afda)
Reviewed-on: https://git-master.nvidia.com/r/2003278
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>